+++ src/libs/kdt.rs
@@ -14,7 +14,7 @@ pub struct KDTree<T> where T: Dimensions {
 
 #[derive(Clone,Copy)]
 enum Tree {
-    Fork( f32   // Dividing line
+    Fork( f64   // Dividing line
         , usize // Index to left tree
         , usize // Index to mid tree
         , usize // Index to right tree
@@ -25,28 +25,28 @@ enum Tree {
 pub trait Dimensions {
     fn bucket_size() -> usize;
     fn num_dims() -> usize;
-    fn dimensions(&self, dim: usize) -> f32;
-    fn radii(&self, dim: usize) -> f32;
+    fn dimensions(&self, dim: usize) -> f64;
+    fn radii(&self, dim: usize) -> f64;
 }
 
 impl<T: Clone + Dimensions> KDTree<T> {
 
     pub fn new(vec: Vec<T>) -> KDTree<T> {
         let len = vec.len();
-        let depth = (len as f32 / <T as Dimensions>::bucket_size() as f32).ceil().log(2.0) as usize;
+        let depth = (len as f64 / <T as Dimensions>::bucket_size() as f64).ceil().log(2.0) as usize;
         let mut kdt = KDTree{trees: [Tree::Leaf(0,0); 1024], vec: vec};
         let (_,tree) = kdt.make_tree(depth, 0, 0, len, 0);
         kdt.trees[0] = tree;
         kdt
     }
 
-    pub fn in_range(&self, pred: &Fn(&T) -> bool, dims: &[(f32,f32)]) -> Vec<T> {
+    pub fn in_range(&self, pred: &Fn(&T) -> bool, dims: &[(f64,f64)]) -> Vec<T> {
         let mut vec = Vec::with_capacity(128);
         KDTree::in_range_matching(self, self.trees[0], pred, dims, 0, &mut vec);
         vec
     }
 
-    pub fn in_range_buff(&self, pred: &Fn(&T) -> bool, dims: &[(f32,f32)], vec: &mut Vec<T>) {
+    pub fn in_range_buff(&self, pred: &Fn(&T) -> bool, dims: &[(f64,f64)], vec: &mut Vec<T>) {
         vec.clear();
         KDTree::in_range_matching(self, self.trees[0], pred, dims, 0, vec);
     }
@@ -83,15 +83,15 @@ impl<T: Clone + Dimensions> KDTree<T> {
         }
     }
 
-    fn mean(dim: usize, vec: &[T], ix: usize, len: usize) -> f32 {
+    fn mean(dim: usize, vec: &[T], ix: usize, len: usize) -> f64 {
         let mut acc = 0.0;
         for i in ix..ix + len {
             acc += vec[i].dimensions(dim);
         }
-        acc / (len as f32)
+        acc / (len as f64)
     }
 
-    fn left_divide(dim: usize, avg: f32, vec: &mut [T], ix: usize, len: usize) -> usize {
+    fn left_divide(dim: usize, avg: f64, vec: &mut [T], ix: usize, len: usize) -> usize {
         let mut c = ix;
         for i in ix..ix + len {
             let e = vec[i].clone();
@@ -105,7 +105,7 @@ impl<T: Clone + Dimensions> KDTree<T> {
     }
 
     // Move all elements who cross the median/avg line to the left side of the slice
-    fn mid_divide(dim: usize, avg: f32, vec: &mut Vec<T>, ix: usize, len: usize) -> usize {
+    fn mid_divide(dim: usize, avg: f64, vec: &mut Vec<T>, ix: usize, len: usize) -> usize {
         let mut c = ix;
         for i in ix..ix + len {
             let e = vec[i].clone();
@@ -118,7 +118,7 @@ impl<T: Clone + Dimensions> KDTree<T> {
         c - ix
     }
 
-    fn in_range_matching(&self, tree: Tree, pred: &Fn(&T) -> bool, dims: &[(f32,f32)], dim: usize, vec: &mut Vec<T>) -> () {
+    fn in_range_matching(&self, tree: Tree, pred: &Fn(&T) -> bool, dims: &[(f64,f64)], dim: usize, vec: &mut Vec<T>) -> () {
         let next_dim = (dim + 1).rem(<T as Dimensions>::num_dims());
         let (crd,rad) = dims[dim];
 
@@ -149,10 +149,10 @@ impl<T: Clone + Dimensions> KDTree<T> {
 pub struct PointAndRadii {
     pub id:         usize,
     pub team:       usize,
-    pub x:          f32,
-    pub y:          f32,
-    pub radius:     f32,
-    pub weight:     f32,
+    pub x:          f64,
+    pub y:          f64,
+    pub radius:     f64,
+    pub weight:     f64,
     pub flying:     bool,
     pub structure:  bool,
     pub missile:    bool,
@@ -162,13 +162,13 @@ pub struct PointAndRadii {
 impl Dimensions for PointAndRadii {
     fn bucket_size() -> usize {512}
     fn num_dims() -> usize {2}
-    fn dimensions(&self, dim: usize) -> f32 {
+    fn dimensions(&self, dim: usize) -> f64 {
         match dim {
             0 => { self.x }
             _ => { self.y }
         }
     }
-    fn radii(&self, _: usize) -> f32 {
+    fn radii(&self, _: usize) -> f64 {
         self.radius
     }
 }
@@ -243,10 +243,10 @@ pub fn bench() {
     }
 
     let build_time = start1.to(end1).num_nanoseconds().unwrap();
-    println!("Build time: {}ms", build_time as f32 / mili);
-    println!("KDT search time: {}ms", total_kdt_search_time as f32 / mili);
-    println!("Naive search time: {}ms", total_search_time as f32 / mili);
-    println!("Improvement: {}", total_search_time as f32 / (total_kdt_search_time + build_time) as f32);
+    println!("Build time: {}ms", build_time as f64 / mili);
+    println!("KDT search time: {}ms", total_kdt_search_time as f64 / mili);
+    println!("Naive search time: {}ms", total_search_time as f64 / mili);
+    println!("Improvement: {}", total_search_time as f64 / (total_kdt_search_time + build_time) as f64);
     println!("KDTree in range: {}", total_in_rng1);
     println!("Naives in range: {} \n", total_in_rng2);
 }
