+++ src/libs/movement.rs
@@ -1,13 +1,13 @@
-use std::f32;
-use std::f32::consts::{PI};
+use std::f64;
+use std::f64::consts::{PI};
 use std::ops::{Add,Sub};
 
 // Area of hexagon
 // 2.5980762113533159402911695122588 * r^2
 
-pub type Point = (f32,f32);
+pub type Point = (f64,f64);
 
-pub fn dist_to_stop(mut speed: f32, deceleration: f32) -> f32 {
+pub fn dist_to_stop(mut speed: f64, deceleration: f64) -> f64 {
     let mut c = 0.0;
     while speed > 0.0 {
         c += speed;
@@ -17,7 +17,7 @@ pub fn dist_to_stop(mut speed: f32, deceleration: f32) -> f32 {
 }
 
 pub trait Collider {
-    fn x_y_radius_weight(&self) -> (f32,f32,f32,f32);
+    fn x_y_radius_weight(&self) -> (f64,f64,f64,f64);
 }
 
 pub fn collide<A: Collider>(a: &A, vec: &[A]) -> Point {
@@ -29,20 +29,20 @@ pub fn collide<A: Collider>(a: &A, vec: &[A]) -> Point {
         let (bx,by,br,bw) = b.x_y_radius_weight();
         let x_dif = ax - bx;
         let y_dif = ay - by;
-        let r_dif = (ar + br) - f32::sqrt(x_dif * x_dif + y_dif * y_dif);
+        let r_dif = (ar + br) - f64::sqrt(x_dif * x_dif + y_dif * y_dif);
         let w_dif = bw / aw;
-        let angl  = f32::atan2(y_dif, x_dif);
-        xo += f32::cos(angl) * r_dif * w_dif;
-        yo += f32::sin(angl) * r_dif * w_dif;
+        let angl  = f64::atan2(y_dif, x_dif);
+        xo += f64::cos(angl) * r_dif * w_dif;
+        yo += f64::sin(angl) * r_dif * w_dif;
     }
     (xo,yo)
 }
 
 #[derive(Clone, Copy, Debug)]
-pub struct Angle(f32);
+pub struct Angle(f64);
 
-pub fn new(x: f32, y: f32) -> Angle {
-    let ang = f32::atan2(y, x);
+pub fn new(x: f64, y: f64) -> Angle {
+    let ang = f64::atan2(y, x);
     if ang < 0.0 {
         Angle(ang + PI * 2.0)
     }
@@ -51,11 +51,11 @@ pub fn new(x: f32, y: f32) -> Angle {
     }
 }
 
-pub unsafe fn make_from(f: f32) -> Angle {
+pub unsafe fn make_from(f: f64) -> Angle {
     Angle(f)
 }
 
-pub fn normalize(mut f: f32) -> Angle {
+pub fn normalize(mut f: f64) -> Angle {
     while f > PI * 2.0 {
         f -= PI * 2.0;
     }
@@ -65,9 +65,9 @@ pub fn normalize(mut f: f32) -> Angle {
     Angle(f)
 }
 
-pub fn denormalize(Angle(f): Angle) -> f32 { f }
+pub fn denormalize(Angle(f): Angle) -> f64 { f }
 
-pub fn distance(Angle(a): Angle, Angle(b): Angle) -> f32 {
+pub fn distance(Angle(a): Angle, Angle(b): Angle) -> f64 {
     let dists = (a - b).abs();
 
     if dists > PI {
@@ -79,7 +79,7 @@ pub fn distance(Angle(a): Angle, Angle(b): Angle) -> f32 {
 }
 
 // Angle to turn, angle to turn towards, amount to turn
-pub fn turn_towards(start: Angle, goal: Angle, turn: f32) -> Angle {
+pub fn turn_towards(start: Angle, goal: Angle, turn: f64) -> Angle {
     let Angle(a) = start;
     let Angle(b) = goal;
     let dist = distance(start, goal);
@@ -115,8 +115,8 @@ pub fn lock_angle(lock: Angle, org: Angle, Angle(arc): Angle) -> Angle {
     }
 }
 
-pub fn move_in_direction(x: f32, y: f32, speed: f32, Angle(ang): Angle) -> Point {
-    (x + f32::cos(ang) * speed, y + f32::sin(ang) * speed)
+pub fn move_in_direction(x: f64, y: f64, speed: f64, Angle(ang): Angle) -> Point {
+    (x + f64::cos(ang) * speed, y + f64::sin(ang) * speed)
 }
 
 impl Add for Angle {
@@ -138,18 +138,18 @@ impl Sub for Angle {
 }
 
 pub fn get_offset_position((x,y): Point, Angle(angle): Angle, (x_off, y_off): Point) -> Point {
-    let coeff = f32::cos(angle);
+    let coeff = f64::cos(angle);
 
     (x + coeff * x_off, y + coeff * y_off)
 }
 
-pub fn circle_line_intersection((ax,ay): Point, (bx,by): Point, (cx,cy): Point, radius: f32) -> Option<Point> {
+pub fn circle_line_intersection((ax,ay): Point, (bx,by): Point, (cx,cy): Point, radius: f64) -> Option<Point> {
 
     if point_in_circle((ax,ay),(cx,cy),radius) {
         return Some((ax,ay));
     }
 
-    if (ax - bx).abs() < f32::EPSILON && (ay - by).abs() < f32::EPSILON {
+    if (ax - bx).abs() < f64::EPSILON && (ay - by).abs() < f64::EPSILON {
         return None;
     }
 
@@ -172,7 +172,7 @@ pub fn circle_line_intersection((ax,ay): Point, (bx,by): Point, (cx,cy): Point, 
         return None;
     }
 
-    let sqrt = f32::sqrt(disc);
+    let sqrt = f64::sqrt(disc);
     let ab_scaling1 = -cross_ratio + sqrt;
     let ab_scaling2 = -cross_ratio + sqrt;
 
@@ -192,14 +192,14 @@ pub fn circle_line_intersection((ax,ay): Point, (bx,by): Point, (cx,cy): Point, 
     }
 }
 
-fn dist_between_points_sqrd((ax,ay): Point, (bx,by): Point) -> f32 {
+fn dist_between_points_sqrd((ax,ay): Point, (bx,by): Point) -> f64 {
     let dx = ax - bx;
     let dy = ay - by;
 
     dx * dx + dy * dy
 }
 
-fn point_in_circle((ax,ay): Point, (bx,by): Point, r: f32) -> bool {
+fn point_in_circle((ax,ay): Point, (bx,by): Point, r: f64) -> bool {
     let dx = ax - bx;
     let dy = ay - by;
 
@@ -211,13 +211,13 @@ pub fn test_circle_line_intersection() {
     println!("{:?}", circle_line_intersection((0.0,0.0), (1.5,0.0), (2.0,0.0), 1.0));
 }
 
-pub fn intercept_point((ax,ay): Point, (bx,by): Point, (vx,vy): Point, s: f32) -> Option<Point> {
+pub fn intercept_point((ax,ay): Point, (bx,by): Point, (vx,vy): Point, s: f64) -> Option<Point> {
     let ox = ax - bx;
     let oy = ay - by;
 
     let h1 = vx * vx + vy * vy - s * s;
     let h2 = ox * vx + oy * vy;
-    let t: f32;
+    let t: f64;
 
     if h1 == 0.0 { // problem collapses into a simple linear equation
         t = -(ox * ox + oy * oy) / (2.0 * h2);
@@ -229,13 +229,13 @@ pub fn intercept_point((ax,ay): Point, (bx,by): Point, (vx,vy): Point, s: f32) -
             return None;
         }
 
-        let root = f32::sqrt(discriminant);
+        let root = f64::sqrt(discriminant);
 
         let t1 = minus_p_half + root;
         let t2 = minus_p_half - root;
 
-        let t_min = f32::min(t1, t2);
-        let t_max = f32::max(t1, t2);
+        let t_min = f64::min(t1, t2);
+        let t_max = f64::max(t1, t2);
 
         t = if t_min > 0.0 { t_min } else { t_max }; // get the smaller of the two times, unless it's negative
